backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
pyjwt
gunicorn
